image generator based on preference input
Used pytorch libraries to train a model and generate images on a ridiculously small dataset. Output was a ridiculous set of pixels LOL. Super fun to actually execute a model, "refine" it by exploring how tweaking parameters such as batch size and features in the generator and discriminator impact performance, and seeing "images" as output.